- Moved octree output functionality to seperate header file (octree_io.h) - Moved performance timers inside octree code, to seperate them from actual writing functions - Fixes for -Wall compiler warnings
