+++ src/ooc_svo_builder/svo_builder/svo_builder.vcxproj.filters
@@ -65,5 +65,8 @@
     <ClInclude Include="Node.h">
       <Filter>Header Files\util</Filter>
     </ClInclude>
+    <ClInclude Include="octree_io.h">
+      <Filter>Header Files\util</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
